+++ src/IML.mli
@@ -4,7 +4,7 @@
 (*                                                                        *)
 (*                       Jun Furuse, DaiLambda, Inc.                      *)
 (*                                                                        *)
-(*                     Copyright 2019  DaiLambda, Inc.                    *)
+(*                   Copyright 2019,2020  DaiLambda, Inc.                 *)
 (*                                                                        *)
 (*   All rights reserved.  This file is distributed under the terms of    *)
 (*   the GNU Lesser General Public License version 2.1, with the          *)
@@ -33,7 +33,9 @@ module PatVar : sig
 end
 
 module Attr : sig
-  type t = Comment of string
+  type t = 
+    | Comment of string
+
   type ts = t list
       
   val add : t -> ('a, ts) with_loc_and_type -> ('a, ts) with_loc_and_type