+++ src/michelson.ml
@@ -4,7 +4,7 @@
 (*                                                                        *)
 (*                       Jun Furuse, DaiLambda, Inc.                      *)
 (*                                                                        *)
-(*                     Copyright 2019  DaiLambda, Inc.                    *)
+(*                   Copyright 2019,2020  DaiLambda, Inc.                 *)
 (*                                                                        *)
 (*   All rights reserved.  This file is distributed under the terms of    *)
 (*   the GNU Lesser General Public License version 2.1, with the          *)
@@ -85,7 +85,9 @@ module Type = struct
     | TyOperation
     | TyContract of t
     | TyLambda of t * t
-  
+
+  let attribute ss t = { t with attrs= t.attrs @ ss }
+
   let mk desc = { desc ; attrs= [] }
 
   let tyString              = mk TyString
@@ -144,20 +146,98 @@ module Type = struct
 
   and pp fmt t = Mline.pp fmt & to_micheline t
 
-  let rec storable ty = match ty.desc with
-    | TyContract _ | TyOperation | TyBigMap _ -> false
+  let rec validate ty = 
+    let open Result.Infix in
+    let rec f ty = match ty.desc with
+      | TyBigMap (k, v) ->
+          f k >>= fun () -> f v >>= fun () ->
+          if not (is_comparable k) then
+            Error (ty, "big_map's key type must be comparable")
+          else if not (is_packable v) then
+            Error (ty, "big_map's value type must be packable")
+          else 
+            Ok ()
+            
+      | TySet e ->
+          f e >>= fun () ->
+          if not (is_comparable e) then 
+            Error (ty, "set's element type must be comparable")
+          else
+            Ok ()
 
-    | TyLambda (_t1, _t2) -> true (* XXX I beieve. (i.e. not sure) *)
+      | TyMap (k, v) ->
+          f k >>= fun () -> f v >>= fun () ->
+          if not (is_comparable k) then
+            Error (ty, "map's key type must be comparable")
+          else 
+            Ok ()
+         
+      | TyContract p ->
+          f p >>= fun () ->
+          if not (is_parameterable p) then
+            Error (ty, "contract's parameter type cannot contain operation")
+          else
+            Ok ()
+          
+      | (TyList ty | TyOption ty) -> f ty
+      | (TyPair (ty1, ty2) | TyOr (ty1, ty2) | TyLambda (ty1, ty2)) -> 
+          f ty1 >>= fun () -> f ty2
 
-    | TyList t | TyOption t | TySet t -> storable t
+      | TyString | TyNat | TyInt | TyBytes | TyBool | TyUnit | TyMutez
+      | TyKeyHash | TyTimestamp | TyAddress |TyChainID | TyKey | TySignature
+      | TyOperation -> Ok ()
+    in
+    f ty
+      
+  and is_comparable ty = 
+    let rec f ty = match ty.desc with
+      | TyBigMap _ | TyChainID | TyContract _ | TyKey 
+      | TyLambda _ | TyList _ | TyMap _ | TyOperation
+      | TyOption _ | TyOr _ | TySet _ | TySignature | TyUnit -> false
+        
+      | TyString | TyNat | TyInt | TyBytes | TyBool | TyMutez 
+      | TyKeyHash | TyTimestamp | TyAddress -> true
+        
+      | TyPair (ty1, ty2) -> f ty1 && f ty2 (* since 005_Babylon *)
+    in
+    f ty
+      
+  and is_packable ty =
+    (* ~allow_big_map:false
+       ~allow_operation:false
+       ~allow_contract:legacy
+    *)
+    let rec f ty = match ty.desc with
+      | TyBigMap _ -> false
+      | TyOperation -> false
+      | TyContract _ -> false
+      | TyLambda (_t1, _t2) -> true
+      | TyList t | TyOption t | TySet t -> f t
+      | TyPair (t1, t2) | TyOr (t1, t2) | TyMap (t1, t2) -> f t1 && f t2
+      | TyString | TyNat | TyInt | TyBytes | TyBool | TyUnit
+      | TyMutez | TyKeyHash | TyTimestamp | TyAddress | TyChainID
+      | TyKey | TySignature -> true
+    in
+    f ty
+      
+  and is_parameterable ty = 
+      (* ~allow_big_map:true
+         ~allow_operation:false
+         ~allow_contract:true
+      *)
+    let rec f ty = match ty.desc with
+      | TyBigMap _ -> true
+      | TyOperation -> false
+      | TyContract _ -> true
 
-    | TyPair (t1, t2) | TyOr (t1, t2)
-    | TyMap (t1, t2) -> storable t1 && storable t2
-
-    | TyString | TyNat | TyInt | TyBytes | TyBool | TyUnit
-    | TyMutez | TyKeyHash | TyTimestamp | TyAddress | TyChainID
-    | TyKey | TySignature -> true
-
+      | TyList t | TyOption t | TySet t -> f t
+      | TyLambda (_t1, _t2) -> true
+      | TyPair (t1, t2) | TyOr (t1, t2) | TyMap (t1, t2) -> f t1 && f t2
+      | TyString | TyNat | TyInt | TyBytes | TyBool | TyUnit
+      | TyMutez | TyKeyHash | TyTimestamp | TyAddress | TyChainID
+      | TyKey | TySignature -> true
+    in
+    f ty
 end
 
 module rec Constant : sig
@@ -310,6 +390,7 @@ and Opcode : sig
   val pp : Format.formatter -> t -> unit
   val to_micheline : t -> Mline.t
   val clean_failwith : t list -> t list
+  val dip_1_drop_n_compaction : t list -> t list
 end = struct
 
   type module_ = 
@@ -444,8 +525,8 @@ end = struct
   
       | EXEC -> !"EXEC"
       | FAILWITH -> !"FAILWITH"
-      | COMMENT (s, ts) ->
-          add_comment (Some s) & seq (List.map f ts)
+      | COMMENT (s, [t]) -> add_comment (Some s) & f t
+      | COMMENT (s, ts) -> add_comment (Some s) & seq (List.map f ts)
       | IF_LEFT (t1, t2) ->
           prim "IF_LEFT" [ seq & List.map f t1;
                            seq & List.map f t2 ]
@@ -579,6 +660,48 @@ end = struct
     | Left t -> Left (constant t)
     | Right t -> Right (constant t)
     | (Unit | Bool _ | Int _ | String _ | Bytes _ | Timestamp _ | Option None as c) -> c
+
+  let dip_1_drop_n_compaction ts =
+    let rec loop n comments = function
+      | DIP (1, [DROP m]) :: ts -> loop (n + m) comments ts
+      | COMMENT (c, [DIP (1, [DROP m])]) :: ts -> loop (n + m) (c :: comments) ts
+      | ts when n > 0 -> 
+          if comments <> [] then
+            COMMENT (String.concat ", " (List.rev comments),
+                     [ DIP (1, [DROP n]) ]) :: loop 0 [] ts
+          else
+            DIP (1, [DROP n]) :: loop 0 [] ts
+      | [] -> []
+      | t :: ts -> 
+          let t' = match t with
+            | DIP (n, ts) -> DIP (n, loop 0 [] ts)
+            | LAMBDA (t1, t2, ts) -> LAMBDA (t1, t2, loop 0 [] ts)
+            | IF (ts1, ts2) -> IF (loop 0 [] ts1, loop 0 [] ts2)
+            | IF_NONE (ts1, ts2) -> IF_NONE (loop 0 [] ts1, loop 0 [] ts2)
+            | IF_LEFT (ts1, ts2) -> IF_LEFT (loop 0 [] ts1, loop 0 [] ts2)
+            | IF_CONS (ts1, ts2) -> IF_CONS (loop 0 [] ts1, loop 0 [] ts2)
+            | COMMENT (c, ts) -> COMMENT (c, loop 0 [] ts)
+            | ITER ts -> ITER (loop 0 [] ts)
+            | MAP ts -> ITER (loop 0 [] ts)
+            | LOOP ts -> LOOP (loop 0 [] ts)
+            | LOOP_LEFT ts -> LOOP_LEFT (loop 0 [] ts)
+
+            | DUP | DIG _ | DUG _ | DROP _ | SWAP | PAIR | ASSERT | CAR | CDR
+            | LEFT _ | RIGHT _ | APPLY | PUSH _ | NIL _ | CONS | NONE _
+            | SOME | COMPARE | EQ | LT | LE | GT | GE | NEQ
+            | ADD | SUB | MUL | EDIV | ABS | ISNAT | NEG | LSL | LSR 
+            | AND | OR | XOR | NOT | EXEC | FAILWITH | UNIT
+            | EMPTY_SET _ | EMPTY_MAP _ | EMPTY_BIG_MAP _
+            | SIZE | MEM | UPDATE | CONCAT | SELF | GET
+            | RENAME _ | PACK | UNPACK _ | SLICE | CAST 
+            | CONTRACT _ | TRANSFER_TOKENS | SET_DELEGATE | CREATE_ACCOUNT
+            | CREATE_CONTRACT _ | IMPLICIT_ACCOUNT | NOW | AMOUNT | BALANCE
+            | CHECK_SIGNATURE | BLAKE2B | SHA256 | SHA512 | HASH_KEY | STEPS_TO_QUOTA
+            | SOURCE | SENDER | ADDRESS | CHAIN_ID -> t
+          in
+          t' :: loop 0 [] ts
+    in
+    loop 0 [] ts
 end
 
 module Module = struct