+++ src/primitives.mli
@@ -4,7 +4,7 @@
 (*                                                                        *)
 (*                       Jun Furuse, DaiLambda, Inc.                      *)
 (*                                                                        *)
-(*                     Copyright 2019  DaiLambda, Inc.                    *)
+(*                   Copyright 2019,2020  DaiLambda, Inc.                 *)
 (*                                                                        *)
 (*   All rights reserved.  This file is distributed under the terms of    *)
 (*   the GNU Lesser General Public License version 2.1, with the          *)
@@ -14,4 +14,8 @@
 
 open Michelson
 
-val primitives : (string * (int * (Type.t -> Opcode.t list -> Opcode.t list))) list
+val primitives : 
+  (string * 
+   (int 
+    * (loc:Location.t -> Type.t -> Opcode.t list -> Opcode.t list)))
+    list